+++ apps/backoffice-v2/src/domains/customer/fetchers.ts
@@ -3,19 +3,27 @@ import { apiClient } from '../../common/api-client/api-client';
 import { Method } from '../../common/enums';
 import { handleZodError } from '../../common/utils/handle-zod-error/handle-zod-error';
 
-const CustomerSchema = z
-  .object({
-    id: z.string(),
-    name: z.string(),
-    displayName: z.string(),
-    logoImageUri: z.union([z.string(), z.null()]).optional(),
-    // Remove default once data migration is done
-    faviconImageUri: z.string().default(''),
-    customerStatus: z.string().optional(),
-    country: z.union([z.string(), z.null()]).optional(),
-    language: z.union([z.string(), z.null()]).optional(),
-  })
-  .optional();
+const CustomerSchema = z.object({
+  id: z.string(),
+  name: z.string(),
+  displayName: z.string(),
+  logoImageUri: z.union([z.string(), z.null()]).optional(),
+  // Remove default once data migration is done
+  faviconImageUri: z.string().default(''),
+  customerStatus: z.string().optional(),
+  country: z.union([z.string(), z.null()]).optional(),
+  language: z.union([z.string(), z.null()]).optional(),
+  config: z
+    .object({
+      isMerchantMonitoringEnabled: z.boolean().default(false),
+      isExample: z.boolean().default(false),
+    })
+    .nullable()
+    .default({
+      isMerchantMonitoringEnabled: false,
+      isExample: false,
+    }),
+});
 
 export type TCustomer = z.infer<typeof CustomerSchema>;