+++ services/workflows-service/prisma/migrations/20240519135000_added_position_in_business_column_to_end_users_on_business/migration.sql
@@ -1,0 +1,5 @@
+-- CreateEnum
+CREATE TYPE "BusinessPosition" AS ENUM ('ubo', 'director', 'representative', 'authorized_signatory');
+
+-- AlterTable
+ALTER TABLE "EndUsersOnBusinesses" ADD COLUMN     "position" "BusinessPosition"[];