+++ packages/react-pdf-toolkit/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ballerine/react-pdf-toolkit
 
+## 1.2.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/ui@0.5.3
+
 ## 1.2.2
 
 ### Patch Changes